feat(student-dashboard): show available tasks
