+++ client/src/api/api.ts
@@ -8490,7 +8490,7 @@ export const MentorsApiFp = function(configuration?: Configuration) {
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
-        async getRandomTask(mentorId: number, courseId: number, options?: AxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<object>> {
+        async getRandomTask(mentorId: number, courseId: number, options?: AxiosRequestConfig): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<void>> {
             const localVarAxiosArgs = await localVarAxiosParamCreator.getRandomTask(mentorId, courseId, options);
             return createRequestFunction(localVarAxiosArgs, globalAxios, BASE_PATH, configuration);
         },
@@ -8540,7 +8540,7 @@ export const MentorsApiFactory = function (configuration?: Configuration, basePa
          * @param {*} [options] Override http request option.
          * @throws {RequiredError}
          */
-        getRandomTask(mentorId: number, courseId: number, options?: any): AxiosPromise<object> {
+        getRandomTask(mentorId: number, courseId: number, options?: any): AxiosPromise<void> {
             return localVarFp.getRandomTask(mentorId, courseId, options).then((request) => request(axios, basePath));
         },
     };
